laptop disk encryption

+++ scripts/nix-install.sh
@@ -2,6 +2,8 @@
 
 DISK=/dev/vda
 
+PASSWORD='horse'
+
 echo '[*] Making partitions...'
 parted $DISK -- mklabel gpt
 parted $DISK -- mkpart root ext4 512MB -8GB
@@ -10,23 +12,30 @@ parted $DISK -- mkpart ESP fat32 1MB 512MB
 parted $DISK -- set 3 esp on
 
 echo '[*] Formatting partitions...'
-mkfs.ext4 -L nixos "${DISK}1"
-mkswap -L swap "${DISK}2"
 mkfs.fat -F 32 -n boot "${DISK}3"
+echo -n "${PASSWORD}" | cryptsetup luksFormat --label="NixOS-Encrypted" "${DISK}1" -
+cryptsetup config "${DISK}1" --label "NixOS-Encrypted"
+echo -n "${PASSWORD}" | cryptsetup open "${DISK}1" cryptroot -
+mkfs.ext4 -L "NixOS-Root" /dev/mapper/cryptroot
+mkswap -L "NixOS-Swap" "${DISK}2"
 
 echo '[*] Mounting partitions...'
-mount /dev/disk/by-label/nixos /mnt
+mount /dev/mapper/cryptroot /mnt
 mkdir -p /mnt/boot
-mount -o umask=077 /dev/disk/by-label/boot /mnt/boot
+mount -o umask=077 "${DISK}3" /mnt/boot
+swapon "${DISK}2"
 
 echo '[*] Updating hardware config...'
 mkdir -p /mnt/etc/nixos
-cp -r ~/dotfiles/laptop/* /mnt/etc/nixos
-rm /mnt/etc/nixos/hardware-configuration.nix
+cp -r ~/shared/.dotfiles/laptop/* /mnt/etc/nixos
 nixos-generate-config --root /mnt
 
 echo '[*] Installing NixOS...'
-nixos-install --no-root-passwd
+nixos-install
 
-reboot
+if [ $? -eq 0 ]; then
+  reboot
+else
+  echo '[!] Install failed.'
+fi
